requirements.txt
sherlock-project
python-nmap
Pillow
exifread
wcwidth
requests
